+++ stock-managment-app/database/migrations/2021_08_04_161229_create_clients_table.php
@@ -16,7 +16,7 @@ class CreateClientsTable extends Migration
         Schema::create('clients', function (Blueprint $table) {
             $table->id();
             $table->string('username');
-            $table->integer('portfolio');
+            $table->float('portfolio');
             $table->timestamps();
         });
     }
